Conflicts arise for the 10.8 binaries, which we had symlinked to the 10.7
binaries, while they seem to have actually compiled new copies.

Conflicts:
	build-support/bin/protobuf/mac/10.8
	build-support/bin/thrift/mac/10.8/0.5.0-finagle
	build-support/bin/thrift/mac/10.8/0.7.0

+++ src/python/twitter/pants/commands/goal.py
@@ -117,6 +117,8 @@ class Goal(Command):
     Option("-l", "--level", dest="log_level", type="choice", choices=['debug', 'info', 'warn'],
            help="[info] Sets the logging level to one of 'debug', 'info' or 'warn', implies -v "
                   "if set."),
+    Option("-n", "--dry-run", action="store_true", dest="dry_run", default=False,
+      help="Print the commands that would be run, without actually running them."),
     Option("--read-from-artifact-cache", "--no-read-from-artifact-cache", action="callback",
       callback=_set_bool, dest="read_from_artifact_cache", default=False,
       help="Whether to read artifacts from cache instead of building them, when possible."),
@@ -172,7 +174,10 @@ class Goal(Command):
   # TODO(John Sirois): revisit wholesale locking when we move py support into pants new
   @classmethod
   def serialized(cls):
-    return True
+    # Goal serialization is now handled in goal execution during group processing.
+    # The goal command doesn't need to hold the serialization lock; individual goals will
+    # acquire the lock if they need to be serialized.
+    return False
 
   def __init__(self, root_dir, parser, args):
     self.targets = []
@@ -370,13 +375,8 @@ class Goal(Command):
       Phase.setup_parser(parser, args, self.phases)
 
   def run(self, lock):
-    with self.check_errors("Target contains a dependency cycle") as error:
-      with self.timer.timing('parse:check_cycles'):
-        for target in self.targets:
-          try:
-            InternalTarget.check_cycles(target)
-          except InternalTarget.CycleException as e:
-            error(target.id)
+    if self.options.dry_run:
+      print '****** Dry Run ******'
 
     logger = None
     if self.options.log or self.options.log_level:
@@ -402,7 +402,13 @@ class Goal(Command):
       for dir in self.options.target_directory:
         self.add_target_directory(dir)
 
-    context = Context(self.config, self.options, self.targets, lock=lock, log=logger)
+    context = Context(
+      self.config,
+      self.options,
+      self.targets,
+      lock=lock,
+      log=logger,
+      timer=self.timer if self.options.time else None)
 
     unknown = []
     for phase in self.phases:
@@ -417,7 +423,7 @@ class Goal(Command):
     if logger:
       logger.debug('Operating on targets: %s', self.targets)
 
-    ret = Phase.attempt(context, self.phases, timer=self.timer if self.options.time else None)
+    ret = Phase.attempt(context, self.phases)
     if self.options.time:
       print('Timing report')
       print('=============')
@@ -435,6 +441,7 @@ class Goal(Command):
 # Install all default pants provided goals
 from twitter.pants.targets import JavaLibrary, JavaTests
 from twitter.pants.tasks.binary_create import BinaryCreate
+from twitter.pants.tasks.build_lint import BuildLint
 from twitter.pants.tasks.bundle_create import BundleCreate
 from twitter.pants.tasks.checkstyle import Checkstyle
 from twitter.pants.tasks.filedeps import FileDeps
@@ -475,9 +482,10 @@ goal(
 
 def async_safe_rmtree(root):
   new_path = root + '.deletable.%f' % time.time()
-  os.rename(root, new_path)
-  with daemon.DaemonContext():
-    safe_rmtree(new_path)
+  if os.path.exists(root):
+    os.rename(root, new_path)
+    with daemon.DaemonContext():
+      safe_rmtree(new_path)
 
 goal(
   name='clean-all-async',
@@ -485,23 +493,25 @@ goal(
   dependencies=['invalidate']
 ).install().with_description('Cleans all intermediate build output in a background process')
 
-if NailgunTask.killall:
-  class NailgunKillall(Task):
-    @classmethod
-    def setup_parser(cls, option_group, args, mkflag):
-      option_group.add_option(mkflag("everywhere"), dest="ng_killall_everywhere",
-                              default=False, action="store_true",
-                              help="[%default] Kill all nailguns servers launched by pants for "
-                                   "all workspaces on the system.")
 
-    def execute(self, targets):
-      if NailgunTask.killall:
-        NailgunTask.killall(self.context.log, everywhere=self.context.options.ng_killall_everywhere)
+class NailgunKillall(Task):
+  @classmethod
+  def setup_parser(cls, option_group, args, mkflag):
+    option_group.add_option(mkflag("everywhere"), dest="ng_killall_everywhere",
+                            default=False, action="store_true",
+                            help="[%default] Kill all nailguns servers launched by pants for "
+                                 "all workspaces on the system.")
 
-  ng_killall = goal(name='ng-killall', action=NailgunKillall)
-  ng_killall.install().with_description('Kill any running nailgun servers spawned by pants.')
+  def execute(self, targets):
+    if NailgunTask.killall:
+      NailgunTask.killall(self.context.log, everywhere=self.context.options.ng_killall_everywhere)
+    else:
+      raise NotImplementedError, 'NailgunKillall not implemented on this platform'
 
-  ng_killall.install('clean-all', first=True)
+ng_killall = goal(name='ng-killall', action=NailgunKillall)
+ng_killall.install().with_description('Kill any running nailgun servers spawned by pants.')
+
+ng_killall.install('clean-all', first=True)
 
 
 # TODO(John Sirois): Resolve eggs
@@ -595,18 +605,39 @@ goal(
   dependencies=['binary']
 ).install().with_description('Create an application bundle from binary targets.')
 
+# run doesn't need the serialization lock. It's reasonable to run some code
+# in a workspace while there's a compile going on unrelated code.
 goal(
   name='jvm-run',
   action=JvmRun,
-  dependencies=['compile']
+  dependencies=['compile'],
+  serialize=False,
 ).install('run').with_description('Run a (currently JVM only) binary target.')
 
 goal(
+  name='jvm-run-dirty',
+  action=JvmRun,
+  serialize=False,
+  ).install('run-dirty').with_description('Run a (currently JVM only) binary target, using\n' +
+    'only currently existing binaries, skipping compilation')
+
+# repl doesn't need the serialization lock. It's reasonable to have
+# a repl running in a workspace while there's a compile going on unrelated code.
+goal(
   name='scala-repl',
   action=ScalaRepl,
-  dependencies=['compile']
+  dependencies=['compile'],
+  serialize=False,
 ).install('repl').with_description(
   'Run a (currently Scala only) REPL with the classpath set according to the targets.')
+
+goal(
+  name='scala-repl-dirty',
+  action=ScalaRepl,
+  serialize=False,
+).install('repl-dirty').with_description(
+  'Run a (currently Scala only) REPL with the classpath set according to the targets, \n' +
+    'using the currently existing binaries, skipping compilation')
 
 goal(
   name='filedeps',
@@ -618,6 +649,13 @@ goal(
   action=PathDeps
 ).install('pathdeps').with_description(
   'Print out a list of all paths containing build files the target depends on')
+
+goal(
+  name='buildlint',
+  action=BuildLint,
+  dependencies=['compile'],  # To pick up missing deps.
+).install()
+
 
 from twitter.pants.tasks.idea_gen import IdeaGen
 
@@ -644,6 +682,7 @@ goal(
   action=Provides,
   dependencies=['jar']
 ).install().with_description('Emit the list of symbols provided by the given targets.')
+
 
 from twitter.pants.tasks.python.setup import SetupPythonEnvironment
